# internal/apps/auth/struct.go
package auth

// Request/Response структуры
type RegisterRequest struct {
	Username string `json:"username"`
	Email    string `json:"email"`
	Password string `json:"password"`
}

type LoginRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type UserResponse struct {
	ID       int    `json:"id"`
	Username string `json:"username"`
	Email    string `json:"email"`
	Admin    int    `json:"admin"`
}

type AuthHandler struct {
	service *AuthService
}